- No need to check for alive senders in a unicast communication
