+++ reactor-aeron/src/main/java/reactor/aeron/processor/UnicastAliveSendersChecker.java
@@ -1,0 +1,24 @@
+package reactor.aeron.processor;
+
+/**
+ * A dummy implementation used in a unicast communication
+ * between an instance of {@link AeronSubscriber} and {@link AeronPublisher}
+ *
+ * @author Anatoly Kadyshev
+ */
+public class UnicastAliveSendersChecker implements AliveSendersChecker {
+
+	@Override
+	public void scheduleCheckForAliveSenders() {
+	}
+
+	@Override
+	public void shutdown() {
+	}
+
+	@Override
+	public boolean isNoSendersDetected() {
+		return false;
+	}
+
+}
